flamenco: fix repair request hdr

+++ src/discof/repair/fd_repair_tile.c
@@ -503,7 +503,7 @@ fd_repair_send_requests( fd_repair_tile_ctx_t * repair_tile_ctx, fd_repair_t * g
         fd_repair_window_index_t * wi = &protocol.inner.window_index;
         fd_hash_copy(&wi->header.sender, glob->public_key);
         fd_hash_copy(&wi->header.recipient, &active->key);
-        wi->header.timestamp = glob->now/1000000L;
+        wi->header.timestamp = (ulong)glob->now/1000000L;
         wi->header.nonce = n;
         wi->slot = ele->dupkey.slot;
         wi->shred_index = ele->dupkey.shred_index;
@@ -517,7 +517,7 @@ fd_repair_send_requests( fd_repair_tile_ctx_t * repair_tile_ctx, fd_repair_t * g
         fd_repair_highest_window_index_t * wi = &protocol.inner.highest_window_index;
         fd_hash_copy(&wi->header.sender, glob->public_key);
         fd_hash_copy(&wi->header.recipient, &active->key);
-        wi->header.timestamp = glob->now/1000000L;
+        wi->header.timestamp = (ulong)glob->now/1000000L;
         wi->header.nonce = n;
         wi->slot = ele->dupkey.slot;
         wi->shred_index = ele->dupkey.shred_index;
@@ -531,7 +531,7 @@ fd_repair_send_requests( fd_repair_tile_ctx_t * repair_tile_ctx, fd_repair_t * g
         fd_repair_orphan_t * wi = &protocol.inner.orphan;
         fd_hash_copy(&wi->header.sender, glob->public_key);
         fd_hash_copy(&wi->header.recipient, &active->key);
-        wi->header.timestamp = glob->now/1000000L;
+        wi->header.timestamp = (ulong)glob->now/1000000L;
         wi->header.nonce = n;
         wi->slot = ele->dupkey.slot;
         FD_LOG_INFO(( "repair request for %lu", ele->dupkey.slot));
